Refactor: move Label to data.models.nearby
