+++ app/src/main/java/fr/free/nrw/commons/nearby/CheckBoxTriStates.java
@@ -7,6 +7,7 @@ import android.widget.CompoundButton;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.AppCompatCheckBox;
 
+import fr.free.nrw.commons.data.models.nearby.Label;
 import java.util.List;
 
 import fr.free.nrw.commons.R;
