+++ app/src/main/java/fr/free/nrw/commons/nearby/NearbyFilterState.java
@@ -1,5 +1,6 @@
 package fr.free.nrw.commons.nearby;
 
+import fr.free.nrw.commons.data.models.nearby.Label;
 import java.util.ArrayList;
 
 public class NearbyFilterState {
