+++ app/src/test/kotlin/fr/free/nrw/commons/ModelFunctions.kt
@@ -3,7 +3,7 @@ import com.nhaarman.mockitokotlin2.whenever
 import fr.free.nrw.commons.Media
 import fr.free.nrw.commons.category.CategoryItem
 import fr.free.nrw.commons.location.LatLng
-import fr.free.nrw.commons.nearby.Label
+import fr.free.nrw.commons.data.models.nearby.Label
 import fr.free.nrw.commons.data.models.nearby.Place
 import fr.free.nrw.commons.data.models.nearby.Sitelinks
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem
